+++ wtforms/ext/appengine/__init__.py
@@ -1,8 +1,8 @@
 import warnings
 
 warnings.warn(
-    'wtforms.ext.appengine is deprecated, and will be removed in WTForms 3.0. '
-    'The package has been split out into its own package, wtforms-appengine: '
-    'https://github.com/wtforms/wtforms-appengine ',
-    DeprecationWarning
+    "'wtforms.ext.appengine' will be removed in WTForms 3.0. Use"
+    " wtforms-appengine instead."
+    " https://github.com/wtforms/wtforms-appengine",
+    DeprecationWarning,
 )