+++ wtforms/ext/django/__init__.py
@@ -1,8 +1,8 @@
 import warnings
 
 warnings.warn(
-    'wtforms.ext.django is deprecated, and will be removed in WTForms 3.0. '
-    'The package has been split out into its own package, wtforms-django: '
-    'https://github.com/wtforms/wtforms-django',
-    DeprecationWarning
+    "'wtforms.ext.django' will be removed in WTForms 3.0. Use"
+    " wtforms-django instead."
+    " https://github.com/wtforms/wtforms-django",
+    DeprecationWarning,
 )